Refactoring into ftp and protocol packages, code optimization.
