+++ hw-project/src/main/java/ru/spbau/mit/ftp/Server.java
@@ -1,4 +1,4 @@
-package ru.spbau.mit;
+package ru.spbau.mit.ftp;
 
 import org.apache.commons.cli.*;
 
@@ -12,12 +12,11 @@ import java.util.concurrent.atomic.AtomicInteger;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
-import ru.spbau.mit.FTPProtocol.Request;
+import ru.spbau.mit.ftp.protocol.*;
 
-public class FTPServer implements Closeable {
-    private static Logger logger = LogManager.getLogger("server");
+public class Server implements Closeable {
+    private static final Logger logger = LogManager.getLogger("server");
     private final ExecutorService executorService;
-    private final File rootDir;
     private volatile boolean interrupt = false;
 
     public static void main(String[] args) {
@@ -25,8 +24,8 @@ public class FTPServer implements Closeable {
         Options options = new Options();
         options.addRequiredOption(null, "port", true, "Port to start listening at");
         options.addOption("j", "threads", true, "number of worker threads to run");
-        int portNumber = 0;
-        int nthreads = 0;
+        int portNumber;
+        int nthreads;
         try {
             logger.info("Parsing options: " + options);
             CommandLine commandLine = parser.parse(options, args);
@@ -40,12 +39,13 @@ public class FTPServer implements Closeable {
         }
 
         logger.info(String.format("Starting server on localhost:%d, with %d threads", portNumber, nthreads));
-        try (FTPServer server = new FTPServer(portNumber, nthreads, new File("/home/kravtsun"))) {
+        try (Server server = new Server(portNumber, nthreads)) {
             Scanner scanner = new Scanner(System.in);
             while (true) {
                 String command = scanner.nextLine();
                 System.out.println("server command: " + command);
                 if (command.equals("exit")) {
+                    server.interrupt = true;
                     break;
                 }
             }
@@ -54,11 +54,7 @@ public class FTPServer implements Closeable {
         }
     }
 
-    public FTPServer(int portNumber, int nthreads, File rootDir) throws FTPServerException {
-        if (!rootDir.isDirectory()) {
-            throw new FTPServerException("invalid root directory: " + rootDir);
-        }
-        this.rootDir = rootDir;
+    public Server(int portNumber, int nthreads) {
         executorService = Executors.newFixedThreadPool(nthreads);
         new Thread(() -> {
             while (!interrupt) {
@@ -74,12 +70,11 @@ public class FTPServer implements Closeable {
 
     @Override
     public void close() throws IOException {
-
     }
 
     private static class FTPServerSession implements Runnable {
-        private static Logger logger = LogManager.getLogger("session");
-        private static AtomicInteger sessionsCount = new AtomicInteger(0);
+        private static final Logger logger = LogManager.getLogger("session");
+        private static final AtomicInteger sessionsCount = new AtomicInteger(0);
         private final Socket socket;
         private final int sessionId;
 
@@ -94,7 +89,7 @@ public class FTPServer implements Closeable {
                     PrintWriter out = new PrintWriter(socket.getOutputStream(), true);
                     BufferedReader in = new BufferedReader(new InputStreamReader(socket.getInputStream()))
             ) {
-                logger.info(logMessage("initializing FTPProtocol"));
+                logger.info(logMessage("initializing Protocol"));
                 out.println("Hello from server, session #" + sessionId);
 
                 logger.info(logMessage("starting IO loop"));
@@ -104,21 +99,21 @@ public class FTPServer implements Closeable {
                         break;
                     }
                     logger.info(logMessage("received: " + inputLine));
-                    FTPProtocol.SentEntity response;
+                    SentEntity response;
                     try {
-                        Request request = FTPProtocol.parseRequest(inputLine);
-                        if (request instanceof FTPProtocol.ListRequest) {
-                            String path = ((FTPProtocol.ListRequest) request).getPath();
+                        Request request = Protocol.parseRequest(inputLine);
+                        if (request instanceof ListRequest) {
+                            String path = ((ListRequest) request).getPath();
                             File[] files = new File(path).listFiles();
-                            response = new FTPProtocol.ListResponse(files);
-//                        } else if (request instanceof FTPProtocol.SimpleRequest) {
-//                            response = new FTPProtocol.SimpleResponse("received request: " + request.requestBody());
+                            response = new ListResponse(files);
+//                        } else if (request instanceof Protocol.SimpleRequest) {
+//                            response = new Protocol.SimpleResponse("received request: " + request.requestBody());
                         } else {
-                            response = new FTPProtocol.SimpleResponse("Unknown request: " + inputLine);
+                            response = new SimpleResponse("Unknown request: " + inputLine);
                         }
-                    } catch (FTPProtocol.FTPProtocolException e) {
+                    } catch (Protocol.FTPProtocolException e) {
                         logger.error(logMessage(e.getMessage()));
-                        response = new FTPProtocol.SimpleResponse("FTPProtocol exception: " + e.getMessage());
+                        response = new SimpleResponse("Protocol exception: " + e.getMessage());
                     }
                     out.println(response.str());
                     logger.info(logMessage("sent: " + response.str()));
@@ -133,12 +128,6 @@ public class FTPServer implements Closeable {
 
         private String logMessage(String message) {
             return String.format("#%d: %s", sessionId, message);
-        }
-    }
-
-    public class FTPServerException extends Exception {
-        FTPServerException(String message) {
-            super(message);
         }
     }
 }
