+++ hw-project/src/main/java/ru/spbau/mit/ftp/protocol/Request.java
@@ -1,0 +1,13 @@
+package ru.spbau.mit.ftp.protocol;
+
+public abstract class Request implements SentEntity {
+
+    @Override
+    public String str() {
+        return code() + " " + requestBody();
+    }
+
+    protected abstract int code();
+
+    protected abstract String requestBody();
+}
