+++ hw-project/src/main/java/ru/spbau/mit/ftp/protocol/SentEntity.java
@@ -1,0 +1,10 @@
+package ru.spbau.mit.ftp.protocol;
+
+public interface SentEntity {
+    String str();
+
+//        default byte []bytes() {
+//            return null;
+//        }
+}
+
