+++ hw-project/src/main/java/ru/spbau/mit/ftp/protocol/SimpleResponse.java
@@ -1,0 +1,14 @@
+package ru.spbau.mit.ftp.protocol;
+
+public class SimpleResponse implements SentEntity {
+    private final String message;
+
+    public SimpleResponse(String message) {
+        this.message = message;
+    }
+
+    @Override
+    public String str() {
+        return message;
+    }
+}
