--- hw-project/src/test/java/ru/spbau/mit/FTPServerTest.java
@@ -1,8 +1,0 @@
-package ru.spbau.mit;
-
-import static org.junit.Assert.*;
-
-// TODO test not only on localhost.
-public class FTPServerTest {
-
-}
